Features/role based authorization

+++ Exoft-BlogWebAPI/Controllers/PostController.cs
@@ -3,64 +3,104 @@ using Business_Logic.Services;
 using Microsoft.AspNetCore.Mvc;
 using Business_Logic.Services.PostServices;
 using Business_Logic.DTO;
+using Microsoft.AspNetCore.Authorization;
+using Business_Logic.Services.UserServices;
 
 namespace Exoft_BlogWebAPI.Controllers
 {
-        [ApiController]
+    [ApiController]
     [Route("api/[controller]")]
     public class PostController : ControllerBase
         {
             IPostService _postService;
+            IAuthService _authService;
 
-
-            public PostController(IPostService postService)
+            public PostController(IPostService postService, IAuthService authService)
             {
                 _postService = postService;
+                _authService = authService;
             }
 
 
             [HttpGet("/posts")]
-            public async Task<IActionResult> GetBlogs()
+            public async Task<IActionResult> GetPosts()
             {
                 return Ok(await _postService.GetAll());
             }
-
+            [HttpGet("/posts-by-user-id/{userId}")]
+            public async Task<IActionResult> GetPostsByUserId(Guid userId)
+            {
+                return Ok(await _postService.GetAllPostsByUserId(userId));
+            }
             [HttpGet("/posts/{id}")]
             public async Task<IActionResult> GetPostById(Guid id)
             {
                 return Ok(await _postService.GetById(id));
             }
 
-
-            [HttpPost]
+            [HttpPost, Authorize]
+            //required return dto with id. 
             public async Task<IActionResult> AddPost(PostCreateDTO post)
-            {
-                
-                await _postService.Create(post);
-                return Ok(post);
-            }
-
-            [HttpPut]
-            public async Task<IActionResult> UpdatePost(PostUpdateDTO post)
-            {
-                await _postService.Update(post);
-                return Ok();
-            }
-
-            [HttpDelete]
-            public IActionResult DeletePost(Guid postId)
             {
                 try
                 {
-                _postService.DeleteById(postId);
-                    return Ok();
+                    var response = await _postService.Create(post);
+                    return Ok(response);
+                } 
+                catch (Exception ex)
+                {
+                    return BadRequest(ex.Message);
+                }
+            }
+
+            [HttpPut, Authorize]
+            public async Task<IActionResult> UpdatePost(PostUpdateDTO post)
+            {
+                if (await _authService.isAuthor(post.UserId))
+                {
+                    var response = await _postService.Update(post);
+                    return Ok(response);
+                } else
+                {
+                    return BadRequest("Can`t Update Post.");
+                }    
+                
+            }
+
+            [HttpDelete, Authorize]
+            public async Task<IActionResult> DeletePost(Guid postId)
+            {
+                try
+                {
+                //problem:  reach the database too many times
+                var post = await _postService.GetById(postId);
+                    if(await _authService.isAuthor(post.UserId))
+                        {
+                            await _postService.DeleteById(postId);
+                            return Ok(new {Status="Post deleted!", PostId=postId });
+                        } else
+                        {
+                            return BadRequest($"Can`t delete post by id: {postId}");
+                        }
+                    
                 }
                 catch (Exception ex)
                 {
                     return BadRequest(ex.Message);
                 }
-
-
+            }
+            [HttpPut("/admin/validate-post"), Authorize(Roles = "Admin")]
+            public async Task<IActionResult> ValidatePost(Guid postId, bool setIsValid)
+            {
+            try
+            {
+                await _postService.ValidatePost(postId, setIsValid);
+                return Ok(new { Response = $"Post is validated.Current valid status: {setIsValid}", ValidStatus = setIsValid });
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(new { Response = $"Can`t validate post." });
+            }
             }
         }
 }
